jenkins: add shared-library boundaries
Introduce explicit shared-library entry points for pipeline execution,
artifact handling, checkout, and the existing pipeline model and
hardware-test helpers while keeping utils.groovy as a compatibility
facade.
