+++ hosts/hetzci/pipeline-library/vars/artifactUtils.groovy
@@ -1,0 +1,99 @@
+// SPDX-FileCopyrightText: 2022-2025 TII (SSRC) and the Ghaf contributors
+// SPDX-License-Identifier: Apache-2.0
+
+def run_cmd(String cmd) {
+  return sh(script: cmd, returnStdout: true).trim()
+}
+
+def run_wget(String url, String to_dir) {
+  sh "wget --show-progress --progress=dot:giga --force-directories --timestamping -P ${to_dir} ${url}"
+  return run_cmd(
+    "wget --force-directories --timestamping -P ${to_dir} ${url} 2>&1 | grep -Po '${to_dir}[^’]+'"
+  )
+}
+
+def path_basename(String path) {
+  if (path == null) {
+    return null
+  }
+  def idx = path.lastIndexOf('/')
+  return idx >= 0 ? path.substring(idx + 1) : path
+}
+
+def image_role(String path) {
+  def basename = path_basename(path)
+  if (basename == null) {
+    return null
+  }
+  return basename.endsWith(".iso") ? "installer" : "disk"
+}
+
+def append_to_build_description(String text) {
+  lock('build-description') {
+    if (!currentBuild.description) {
+      currentBuild.description = text
+    } else {
+      currentBuild.description = "${currentBuild.description}<br>${text}"
+    }
+  }
+}
+
+def append_to_build_description_once(String text) {
+  lock('build-description') {
+    if (currentBuild.description?.contains(text)) {
+      return
+    }
+    if (!currentBuild.description) {
+      currentBuild.description = text
+    } else {
+      currentBuild.description = "${currentBuild.description}<br>${text}"
+    }
+  }
+}
+
+def archive_robot_artifacts(String tmp_img_dir, boolean should_archive) {
+  if (should_archive) {
+    def test_artifacts = '' +
+      'Robot-Framework/test-suites/**/*.html, ' +
+      'Robot-Framework/test-suites/**/*.xml, ' +
+      'Robot-Framework/test-suites/**/*.png, ' +
+      'Robot-Framework/test-suites/**/*.jpeg, ' +
+      'Robot-Framework/test-suites/**/*.mp4, ' +
+      'Robot-Framework/test-suites/**/*.mkv, ' +
+      'Robot-Framework/test-suites/**/*.wav, ' +
+      'Robot-Framework/test-suites/**/*.txt'
+    archiveArtifacts allowEmptyArchive: true, artifacts: test_artifacts
+  }
+  sh "rm -rf ${tmp_img_dir} || true"
+}
+
+def controller_workdir() {
+  def rawTag = env.BUILD_TAG ?: "${env.JOB_NAME}-${env.BUILD_NUMBER}"
+  def tag = pipelineModel.safe_path_component(rawTag)
+  if (!tag || tag == 'null-null') {
+    error('Cannot derive a unique controller workdir for this build')
+  }
+  return "/var/lib/jenkins/ghaf-pipeline-workspaces/${tag}"
+}
+
+def build_tmpdir(String name) {
+  def component = pipelineModel.safe_path_component(name)
+  if (!component || component == 'null') {
+    error("Cannot derive a unique build tmpdir for '${name}'")
+  }
+  return "${controller_workdir()}/tmp/${component}"
+}
+
+def clean_controller_workdir() {
+  node('built-in') {
+    sh "rm -rf '${controller_workdir()}'"
+  }
+}
+
+def with_controller_workspace(String workspace, Closure body) {
+  node('built-in') {
+    dir(workspace) {
+      body()
+    }
+  }
+}
